requirements.txt
backports.functools-lru-cache==1.6.1
certifi==2019.11.28
cloudpickle==1.2.2
cycler==0.10.0
decorator==4.4.1
easydict==1.9
future==0.18.2
kiwisolver==1.1.0
matplotlib==2.2.4
networkx==2.2
numpy==1.16.5
opencv-python==4.1.2.30
pandas==0.24.2
Pillow==6.2.1
protobuf==3.11.1
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.0.3
PyYAML==5.2
scikit-image==0.14.5
scikit-learn==0.20.4
scipy==1.2.2
seaborn==0.9.0
six==1.13.0
subprocess32==3.5.4
tensorboardX==1.9
torch==1.3.1
torchvision==0.4.2
tqdm==4.40.2
